In HTML5 script tags do not need a type attribute.
It's assumed that the type will be JavaScript. This is a
legacy attribute from the days when there was a proposal to
have browsers support multiple kinds of scripts. Now
JavaScript is the only thing that can be in a script tag
so there's no need for a type attribute to be specified.